rename laser plugin files
