+++ src/laser/laser_plugin.h
